Refactor config code
Place config searching and parsing code to own file, config.py
This de-clutters formatter.py and snakefmt.py
